+++ lib/widgets/viewer/video/db_playback_state_handler.dart
@@ -1,10 +1,12 @@
 import 'dart:async';
 
+import 'package:aves/model/settings/settings.dart';
 import 'package:aves/model/video_playback.dart';
 import 'package:aves/services/common/services.dart';
 import 'package:aves/theme/format.dart';
 import 'package:aves/widgets/common/extensions/build_context.dart';
 import 'package:aves/widgets/dialogs/aves_dialog.dart';
+import 'package:aves_model/aves_model.dart';
 import 'package:aves_video/aves_video.dart';
 import 'package:flutter/material.dart';
 
@@ -21,25 +23,32 @@ class DatabasePlaybackStateHandler extends PlaybackStateHandler {
     // clear on retrieval
     await metadataDb.removeVideoPlayback({entryId});
 
-    final resume = await showDialog<bool>(
-      context: context,
-      builder: (context) => AvesDialog(
-        content: Text(context.l10n.videoResumeDialogMessage(formatFriendlyDuration(Duration(milliseconds: resumeTime)))),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.maybeOf(context)?.pop(false),
-            child: Text(context.l10n.videoStartOverButtonLabel),
-          ),
-          TextButton(
-            onPressed: () => Navigator.maybeOf(context)?.pop(true),
-            child: Text(context.l10n.videoResumeButtonLabel),
-          ),
-        ],
-      ),
-      routeSettings: const RouteSettings(name: AvesDialog.confirmationRouteName),
-    );
-    if (resume == null || !resume) return 0;
-    return resumeTime;
+    switch (settings.videoResumptionMode) {
+      case VideoResumptionMode.never:
+        return 0;
+      case VideoResumptionMode.ask:
+        final resume = await showDialog<bool>(
+              context: context,
+              builder: (context) => AvesDialog(
+                content: Text(context.l10n.videoResumeDialogMessage(formatFriendlyDuration(Duration(milliseconds: resumeTime)))),
+                actions: [
+                  TextButton(
+                    onPressed: () => Navigator.maybeOf(context)?.pop(false),
+                    child: Text(context.l10n.videoStartOverButtonLabel),
+                  ),
+                  TextButton(
+                    onPressed: () => Navigator.maybeOf(context)?.pop(true),
+                    child: Text(context.l10n.videoResumeButtonLabel),
+                  ),
+                ],
+              ),
+              routeSettings: const RouteSettings(name: AvesDialog.confirmationRouteName),
+            ) ??
+            false;
+        return resume ? resumeTime : 0;
+      case VideoResumptionMode.always:
+        return resumeTime;
+    }
   }
 
   @override